Convert images to float type
This got lost when removing the normalization.

+++ Exercise/Exercise06/sgm.py
@@ -86,8 +86,8 @@ def preCalculateCosts(imgL, imgR, numDisp):
 	return C
 
 def sgm(imgL, imgR, p1, p2, disparityRange, directions=8):
-	imgL = cv2.cvtColor(imgL, cv2.COLOR_BGR2GRAY)
-	imgR = cv2.cvtColor(imgR, cv2.COLOR_BGR2GRAY)
+	imgL = cv2.cvtColor(imgL, cv2.COLOR_BGR2GRAY).astype(np.float)
+	imgR = cv2.cvtColor(imgR, cv2.COLOR_BGR2GRAY).astype(np.float)
 
 	dispRange = range(disparityRange[0], disparityRange[1])
 	numDisp = len(dispRange)
